Now Alea only replies to messages aimed at it
It determines a message is aimed at it if the message starts with its
nick or a sigil character, or is in a private message.

I think some of the predicates I wrote need refactoring, though; they
depend on the session and message objects directly, which are
complicated to assemble for testing.